+++ packages/data-schema/__tests__/ModelType.test-d.ts
@@ -102,6 +102,6 @@ describe('identifiers', () => {
       .model({
         title: string(),
       })
-      .secondaryIndexes([a.index('title')]);
+      .secondaryIndexes((index) => [index('title')]);
   });
 });